Made good progress
I've forgotten to push my changes for a bit so I'm pushing now. I've got things in a pretty good place. I'm quite happy with how the game is turning out so far. Next steps are to add UI, audio, and particle effects. I want to add health packs and maybe a weapon powerup, but I need to get the requirements done first.